DC Power
Power input section converted to 24VDC Max input instead of AC. A proper AC to DC power supply would need more components to be properly implmented and not require redesign of the DCDC regulators. Also the plan going forward is to be able to run thes off rechargeable battery packs as well as 24VDC bus voltages.